packages/api/pyproject.toml
[project]
name = "ndif_api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref>=3.10.0",
    "boto3>=1.40.58",
    "eventlet>=0.40.3",
    "fastapi==0.108.0",
    "fastapi-cache2>=0.2.2",
    "fastapi-socketio>=0.0.10",
    "gunicorn>=23.0.0",
    "influxdb-client>=1.49.0",
    "nnsight>=0.5.9",
    "opentelemetry-api>=1.38.0",
    "opentelemetry-exporter-otlp>=1.38.0",
    "opentelemetry-instrumentation-fastapi>=0.59b0",
    "opentelemetry-sdk>=1.38.0",
    "prometheus-client>=0.23.1",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "psycopg2-binary>=2.9.11",
    "python-logging-loki>=0.3.1",
    "python-slugify>=8.0.4",
    "python-socketio==5.13.0",
    "ray[serve]==2.47.0",
    "redis>=7.0.0",
    "requests>=2.32.5",
    "torch>=2.9.0",
    "uvicorn[standard]==0.24.0",
    "ndif_common"
]

[tool.uv.sources]
ndif_common = { workspace = true }
